requirements.txt
torch==2.6.0
torchvision==0.21.0
scikit-learn==1.6.1
pillow==11.0.0
matplotlib==3.10.3
